+++ udata_front/theme/gouvfr/assets/js/components/dataset/CardLG.stories.ts
@@ -1,8 +1,10 @@
-import { CERTIFIED, PUBLIC_SERVICE } from '../../composables/useOrganizationCertified';
+import { CERTIFIED, PUBLIC_SERVICE, type Organization, type User } from '@etalab/data.gouv.fr-components';
 import CardLG from './card-lg.vue';
 import { withActions } from '@storybook/addon-actions/decorator';
+import { Meta, StoryObj } from '@storybook/vue3';
+import { Dataset } from '../../types';
 
-export default {
+const meta = {
   title: 'Components/Dataset Card/Card Large',
   component: CardLG,
   parameters: {
@@ -11,44 +13,68 @@ export default {
     },
   },
   decorators: [withActions],
+} satisfies Meta<typeof CardLG>;
+
+export default meta;
+
+type CardLGProps = {
+  readonly dataset: Dataset,
+  showMetrics?: boolean | undefined,
+  style?: Record<string, any> | undefined,
 };
 
-/**
- * @typedef {{
- *  [Property in keyof import("../../types").NewDataset
- *    as Extract<Property, "acronym" | "archived" | "description" | "last_update" | "license" | "page" | "private" | "quality" | "title" | "owner" | "organization">
- *  ]: import("../../types").NewDataset[Property]
- * } & {metrics?: Object, showMetrics: Boolean, style?: Object}} CardLGProps
- */
+const owner: User = {
+  id: "someId",
+  first_name: "Jean",
+  last_name: "Dupond",
+  page: "http://www.data.gouv.fr",
+};
 
-/** @type {CardLGProps} */
-const args = {
-  acronym: "",
-  archived: false,
-  description: "Some description",
-  last_update: new Date(),
-  license: "lov2",
-  page: "https://www.data.gouv.fr",
-  private: false,
-  quality: {
-    all_resources_available: true,
-    dataset_description_quality: true,
-    has_open_format: true,
-    has_resources: true,
-    license: true,
-    resources_documentation: true,
-    score: 1,
-    spatial: true,
-    temporal_coverage: true,
-    update_frequency: true,
-    update_fulfilled_in_time: true,
+const args: CardLGProps = {
+  dataset: {
+    id: "someId",
+    acronym: "",
+    archived: false,
+    description: "Some description",
+    last_update: (new Date()).toDateString(),
+    license: "lov2",
+    page: "https://www.data.gouv.fr",
+    private: false,
+    quality: {
+      all_resources_available: true,
+      dataset_description_quality: true,
+      has_open_format: true,
+      has_resources: true,
+      license: true,
+      resources_documentation: true,
+      score: 1,
+      spatial: true,
+      temporal_coverage: true,
+      update_frequency: true,
+      update_fulfilled_in_time: true,
+    },
+    title: "My new dataset",
+    resources: [],
+    community_resources: [],
+    created_at: "2014-12-12T12:01:06.889000+00:0",
+    last_modified: "2020-11-03T09:16:55.837000+00:00",
+    uri: "https://www.data.gouv.fr",
+    slug: "data-gouv-fr",
+    tags: [],
+    frequency: "Unknown",
+    temporal_coverage: "Test",
+    frequency_date: null,
+    spatial: null,
+    organization: null,
+    owner,
+    metrics: {
+      discussions: 20,
+      reuses: 100,
+      followers: 500,
+      views: 50000
+    },
   },
-  title: "My new dataset",
   showMetrics: true,
-  metrics: {
-    reuses: 100,
-    favorites: 500
-  }
 };
 
 const updateLastMonth = new Date();
@@ -57,81 +83,88 @@ updateLastMonth.setMonth(updateLastMonth.getMonth() - 1, 20);
 const updateLastYear = new Date();
 updateLastYear.setFullYear(updateLastYear.getFullYear() - 1);
 
-/** @type {import('../../types').User} */
-const owner = {
-  first_name: "Jean",
-  last_name: "Dupond",
-};
-
-/** @type {import('../../types').User} */
-const ownerWithAvatar = {
+const ownerWithAvatar: User = {
   ...owner,
   avatar_thumbnail: "https://static.data.gouv.fr/avatars/0d/aebc1b126b410aa73701fde285dd76-100.jpg"
-}
-
-/** @type {import('../../types').Organization} */
-const organization = {
-  name: "My Organization",
-  badges: [],
-  page: "https://www.data.gouv.fr/fr/organizations/data-gouv-fr/"
 };
 
-/** @type {import('../../types').Organization} */
-const organizationWithLogo = {
+const organization: Organization = {
+  id: "someId",
+  acronym: null,
+  name: "My Organization",
+  badges: [],
+  page: "https://www.data.gouv.fr/fr/organizations/data-gouv-fr/",
+  uri: "https://www.data.gouv.fr/fr/organizations/data-gouv-fr/",
+  slug: "data-gouv-fr",
+  logo: "https://static.data.gouv.fr/avatars/09/1ba932cbfa48dc8c158981de6c700a.jpeg",
+  logo_thumbnail: "https://static.data.gouv.fr/avatars/09/1ba932cbfa48dc8c158981de6c700a-100.jpeg",
+};
+
+const organizationWithLogo: Organization = {
   ...organization,
   logo_thumbnail: "https://static.data.gouv.fr/avatars/09/1ba932cbfa48dc8c158981de6c700a-100.jpeg"
 };
 
-/** @type {CardLGProps} */
-const argsWithOwner = {
+const argsWithOwnerWithAvatar: CardLGProps = {
   ...args,
-  owner,
-};
-
-/** @type {CardLGProps} */
-const argsWithOwnerWithAvatar = {
-  ...args,
-  owner: ownerWithAvatar,
-};
-
-/** @type {CardLGProps} */
-const argsWithOrganization = {
-  ...args,
-  organization: organization,
-};
-
-/** @type {CardLGProps} */
-const argsWithOrganizationWithLogo = {
-  ...args,
-  organization: organizationWithLogo,
-};
-
-/** @type {CardLGProps} */
-const argsWithAcronymAndOrganizationWithLogo = {
-  ...argsWithOrganizationWithLogo,
-  acronym: "MND"
-};
-
-/** @type {CardLGProps} */
-const argsWithAcronymAndOwnerWithAvatar = {
-  ...argsWithOwnerWithAvatar,
-  acronym: "MND"
-};
-
-/** @type {CardLGProps} */
-const argsWithCertifiedOrganization = {
-  ...argsWithAcronymAndOrganizationWithLogo,
-  organization: {
-    ...organizationWithLogo,
-    badges: [{kind: PUBLIC_SERVICE}, {kind: CERTIFIED}]
+  dataset: {
+    ...args.dataset,
+    owner: ownerWithAvatar,
+    organization: null,
   },
+};
+
+const argsWithOrganization: CardLGProps = {
+  ...args,
+  dataset: {
+    ...args.dataset,
+    organization: organization,
+    owner: null,
+  },
+};
+
+const argsWithOrganizationWithLogo: CardLGProps = {
+  ...args,
+  dataset: {
+    ...args.dataset,
+    organization: organizationWithLogo,
+    owner: null,
+  },
+};
+
+const argsWithAcronymAndOrganizationWithLogo: CardLGProps = {
+  ...args,
+  dataset: {
+  ...argsWithOrganizationWithLogo.dataset,
+  acronym: "MND",
+  },
+};
+
+const argsWithAcronymAndOwnerWithAvatar: CardLGProps = {
+  ...args,
+  dataset: {
+    ...argsWithOwnerWithAvatar.dataset,
+    acronym: "MND",
+  },
+};
+
+const argsWithCertifiedOrganization: CardLGProps = {
+  ...args,
+  dataset: {
+    ...argsWithAcronymAndOrganizationWithLogo.dataset,
+    organization: {
+      ...organizationWithLogo,
+      badges: [{kind: PUBLIC_SERVICE}, {kind: CERTIFIED}]
+    },
+    owner: null,
+  }
 };
 
 /**
  * #### Card Large with an organization logo
  * This is how a dataset owned by an organization is shown.
  */
-export const CardLGWithOrganizationLogo = {
+export const CardLGWithOrganizationLogo: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -146,7 +179,7 @@ export const CardLGWithOrganizationLogo = {
  * #### Card Large with a certified organization
  * This is how a dataset owned by a certified organization is shown.
  */
-export const CardLGWithCertifiedOrganization = {
+export const CardLGWithCertifiedOrganization: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -161,7 +194,7 @@ export const CardLGWithCertifiedOrganization = {
  * #### Card Large with an acronym and an organization logo
  * This is how a dataset owned by an organization is shown.
  */
-export const CardLGWithAcronymAndOrganizationLogo = {
+export const CardLGWithAcronymAndOrganizationLogo: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -176,7 +209,7 @@ export const CardLGWithAcronymAndOrganizationLogo = {
  * #### Card Large with a description and a certified organization
  * This is how a dataset with a really long description is shown.
  */
-export const CardLGWithDescriptionAndCertifiedOrganization = {
+export const CardLGWithDescriptionAndCertifiedOrganization: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -186,23 +219,26 @@ export const CardLGWithDescriptionAndCertifiedOrganization = {
   }),
   args: {
     ...argsWithCertifiedOrganization,
-    description: `Le plan cadastral est le découpage du territoire français en unités de surfaces permettant le calcul de certains impôts.
+    dataset: {
+      ...argsWithCertifiedOrganization.dataset,
+      description: `Le plan cadastral est le découpage du territoire français en unités de surfaces permettant le calcul de certains impôts.
 
-    Chaque commune est subdivisée en sections, elles-mêmes subdivisées en feuilles (ou planches).
-    Une feuille cadastrale comporte des parcelles, qui peuvent supporter des bâtiments.
+        Chaque commune est subdivisée en sections, elles-mêmes subdivisées en feuilles (ou planches).
+        Une feuille cadastrale comporte des parcelles, qui peuvent supporter des bâtiments.
 
-    Différence par rapport au PCI Vecteur
-    Contrairement au Plan Cadastral Informatisé au format EDIGÉO qui est un ensemble de 600 000 feuilles cadastrales avec de nombreux éléments liés à la fiscalité ou à l'habillage du plan, cette version retravaillée par Etalab se concentre sur le découpage parcellaire et sur les bâtiments.
+        Différence par rapport au PCI Vecteur
+        Contrairement au Plan Cadastral Informatisé au format EDIGÉO qui est un ensemble de 600 000 feuilles cadastrales avec de nombreux éléments liés à la fiscalité ou à l'habillage du plan, cette version retravaillée par Etalab se concentre sur le découpage parcellaire et sur les bâtiments.
 
-    Les différentes couches d'information géographique sont constituées par juxtaposition des feuilles, sans correction.`
-  },
+        Les différentes couches d'information géographique sont constituées par juxtaposition des feuilles, sans correction.`
+      },
+    }
 };
 
 /**
  * #### Card Large updated last month with an organization logo
  * This is how a dataset owned by an organization is shown.
  */
-export const CardLGUpdatedLastMonthWithOrganizationLogo = {
+export const CardLGUpdatedLastMonthWithOrganizationLogo: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -212,7 +248,10 @@ export const CardLGUpdatedLastMonthWithOrganizationLogo = {
   }),
   args: {
     ...argsWithOrganizationWithLogo,
-    last_update: updateLastMonth
+    dataset: {
+      ...argsWithOrganizationWithLogo.dataset,
+      last_update: updateLastMonth.toISOString()
+    }
   },
 };
 
@@ -220,7 +259,7 @@ export const CardLGUpdatedLastMonthWithOrganizationLogo = {
  * #### Card Large updated last year with an organization logo
  * This is how a dataset owned by an organization is shown.
  */
-export const CardLGUpdatedLastYearWithOrganizationLogo = {
+export const CardLGUpdatedLastYearWithOrganizationLogo: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -230,7 +269,10 @@ export const CardLGUpdatedLastYearWithOrganizationLogo = {
   }),
   args: {
     ...argsWithOrganizationWithLogo,
-    last_update: updateLastYear
+    dataset: {
+      ...argsWithOrganizationWithLogo.dataset,
+      last_update: updateLastYear.toISOString()
+    }
   },
 };
 
@@ -238,7 +280,7 @@ export const CardLGUpdatedLastYearWithOrganizationLogo = {
  * #### Card Large without metrics with an organization logo
  * This is how a dataset without metrics is shown.
  */
-export const CardLGWithoutMetricsWithOrganizationLogo = {
+export const CardLGWithoutMetricsWithOrganizationLogo: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -256,7 +298,7 @@ export const CardLGWithoutMetricsWithOrganizationLogo = {
  * #### Card Large for a private dataset with an organization logo
  * This is how a private dataset owned by an organization is shown.
  */
-export const CardLGPrivateWithOrganizationLogo = {
+export const CardLGPrivateWithOrganizationLogo: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -266,7 +308,10 @@ export const CardLGPrivateWithOrganizationLogo = {
   }),
   args: {
     ...argsWithOrganizationWithLogo,
-    private: true
+    dataset: {
+      ...argsWithOrganizationWithLogo.dataset,
+      private: true,
+    },
   },
 };
 
@@ -274,7 +319,7 @@ export const CardLGPrivateWithOrganizationLogo = {
  * #### Card Large for an archived dataset with organization logo
  * This is how an archived dataset owned by an organization is shown.
  */
-export const CardLGArchivedWithOrganizationLogo = {
+export const CardLGArchivedWithOrganizationLogo: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -284,7 +329,10 @@ export const CardLGArchivedWithOrganizationLogo = {
   }),
   args: {
     ...argsWithOrganizationWithLogo,
-    archived: true
+    dataset: {
+      ...argsWithOrganizationWithLogo.dataset,
+      archived: true,
+    },
   },
 };
 
@@ -292,7 +340,7 @@ export const CardLGArchivedWithOrganizationLogo = {
  * #### Card Large for a private archived dataset with organization logo
  * This is how a private archived dataset owned by an organization is shown.
  */
-export const CardLGPrivateArchivedWithOrganizationLogo = {
+export const CardLGPrivateArchivedWithOrganizationLogo: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -302,8 +350,11 @@ export const CardLGPrivateArchivedWithOrganizationLogo = {
   }),
   args: {
     ...argsWithOrganizationWithLogo,
-    archived: true,
-    private: true,
+    dataset: {
+      ...argsWithOrganizationWithLogo.dataset,
+      archived: true,
+      private: true,
+    },
   },
 };
 
@@ -311,7 +362,7 @@ export const CardLGPrivateArchivedWithOrganizationLogo = {
  * #### Card Large with owner avatar
  * This is how a dataset with an owner is shown.
  */
-export const CardLGWithOwnerAvatar = {
+export const CardLGWithOwnerAvatar: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -326,7 +377,7 @@ export const CardLGWithOwnerAvatar = {
  * #### Card Large with acronym and owner avatar
  * This is how a dataset with an owner is shown.
  */
-export const CardLGWithAcronymAndOwnerWithAvatar = {
+export const CardLGWithAcronymAndOwnerWithAvatar: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -341,7 +392,7 @@ export const CardLGWithAcronymAndOwnerWithAvatar = {
  * #### Card Large for a private dataset with owner avatar
  * This is how a private dataset with an owner is shown.
  */
-export const CardLGPrivateWithOwnerAvatar = {
+export const CardLGPrivateWithOwnerAvatar: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -351,7 +402,10 @@ export const CardLGPrivateWithOwnerAvatar = {
   }),
   args: {
     ...argsWithOwnerWithAvatar,
-    private: true
+    dataset: {
+      ...argsWithOwnerWithAvatar.dataset,
+      private: true,
+    }
   },
 };
 
@@ -359,7 +413,7 @@ export const CardLGPrivateWithOwnerAvatar = {
  * #### Card Large for an archived dataset with owner avatar
  * This is how an archived dataset with an owner is shown.
  */
-export const CardLGArchivedWithOwnerAvatar = {
+export const CardLGArchivedWithOwnerAvatar: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -369,7 +423,10 @@ export const CardLGArchivedWithOwnerAvatar = {
   }),
   args: {
     ...argsWithOwnerWithAvatar,
-    archived: true
+    dataset: {
+      ...argsWithOwnerWithAvatar.dataset,
+      archived: true,
+    }
   },
 };
 
@@ -377,7 +434,7 @@ export const CardLGArchivedWithOwnerAvatar = {
  * #### Card Large for a private archived dataset with owner avatar
  * This is how a private archived dataset with an owner is shown.
  */
-export const CardLGPrivateArchivedWithOwnerAvatar = {
+export const CardLGPrivateArchivedWithOwnerAvatar: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -387,8 +444,11 @@ export const CardLGPrivateArchivedWithOwnerAvatar = {
   }),
   args: {
     ...argsWithOwnerWithAvatar,
-    archived: true,
-    private: true,
+    dataset: {
+      ...argsWithOwnerWithAvatar.dataset,
+      archived: true,
+      private: true,
+    }
   },
 };
 
@@ -396,7 +456,7 @@ export const CardLGPrivateArchivedWithOwnerAvatar = {
  * #### Card Large with organization placeholder
  * This is how a dataset owned by an organization is shown.
  */
-export const CardLGWithOrganizationPlaceholder = {
+export const CardLGWithOrganizationPlaceholder: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -411,23 +471,7 @@ export const CardLGWithOrganizationPlaceholder = {
  * #### Card Large with owner placeholder
  * This is how a dataset with an owner is shown. When the user doesn't have an avatar, udata computes an identicon.
  */
-export const CardLGWithOwnerPlaceholder = {
-  render: (args) => ({
-    components: { CardLG },
-    setup() {
-      return { args };
-    },
-    template: '<CardLG v-bind="args"/>',
-  }),
-  args: argsWithOwner,
-};
-
-/**
- * #### Card Large with dataset placeholder
- * This is how a dataset without owner is shown.
- * It shouldn't exist based on API requirement but it's a `else` path for the card.
- */
-export const CardLGWithDatasetPlaceholder = {
+export const CardLGWithOwnerPlaceholder: StoryObj<typeof meta> = {
   render: (args) => ({
     components: { CardLG },
     setup() {
@@ -436,4 +480,28 @@ export const CardLGWithDatasetPlaceholder = {
     template: '<CardLG v-bind="args"/>',
   }),
   args,
+};
+
+/**
+ * #### Card Large with dataset placeholder
+ * This is how a dataset without owner is shown.
+ * It shouldn't exist based on API requirement but it's a `else` path for the card.
+ */
+export const CardLGWithDatasetPlaceholder: StoryObj<typeof meta> = {
+  render: (args) => ({
+    components: { CardLG },
+    setup() {
+      return { args };
+    },
+    template: '<CardLG v-bind="args"/>',
+  }),
+  args: {
+    ...args,
+    // @ts-ignore
+    dataset: {
+      ...args.dataset,
+      owner: null,
+      organization: null,
+    }
+  },
 };