+++ udata_front/theme/gouvfr/datagouv-components/src/components/CopyButton/CopyButton.vue
@@ -15,7 +15,10 @@ import { useI18n } from "vue-i18n";
 import Clipboard from "clipboard";
 
 defineProps<{
-  text: string
+  /**
+   * The text to copy
+   */
+  text: string,
 }>();
 
 const { t } = useI18n();