+++ udata_front/theme/gouvfr/datagouv-components/src/components/CopyButton/index.ts
@@ -1,0 +1,5 @@
+import "../base.less";
+import "../button.less";
+import CopyButton from "./CopyButton.vue";
+
+export { CopyButton };