+++ udata_front/theme/gouvfr/datagouv-components/src/types/users.ts
@@ -4,6 +4,7 @@ export type User = {
   first_name: string;
   last_name: string;
   avatar?: string;
+  page: string;
   avatar_thumbnail?: string;
   roles?: Array<string>;
 }